+++ src/components/manage/Contents/ContentsPropertiesModal.jsx
@@ -205,7 +205,6 @@ class ContentsPropertiesModal extends Component {
                 choices: [
                   [true, 'Yes'],
                   [false, 'No'],
-                  [null, 'Leave settings as they are'],
                 ],
               },
             },