+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Feature
 
+- added loading icon when doing actions in folder-contents @giuliaghisini
+
 ### Bugfix
 
 ### Internal